+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * [Ecto.Schema] Support :returning option for delete
 
+### Bug fixes
+
+* [EctoQLC.Adapters.QLC] Ensure record elements are updated correctly
+
 ## Ecto QLC v0.1.0 December 16th, 2022
 
 ### Enhancements